+++ app/src/main/res/drawable/repeat_onoff.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_pressed="true" android:color="@color/red"/>
-    <item android:state_pressed="false" android:color="@color/gray2"/>
+    <item android:state_checked="true" android:color="@color/red"/>
+    <item android:state_checked="false" android:color="@color/gray2"/>
 </selector>
