+++ app/src/main/res/layout/activity_intro.xml
@@ -5,7 +5,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="?attr/fullscreenBackgroundColor"
-        tools:context=".IntroActivity">
+        tools:context=".intro.IntroActivity">
 
     <!-- The primary full-screen view. This can be replaced with whatever view
          is needed to present your content, e.g. VideoView, SurfaceView,
