+++ app/src/main/res/layout/activity_places.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".PlacesActivity">
+    tools:context=".place.PlacesActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -13,9 +13,10 @@
         android:layout_marginBottom="20dp"
         android:orientation="horizontal"
         android:layout_marginTop="15dp"
-        tools:context=".PlacesActivity">
+        tools:context=".place.PlacesActivity">
 
         <ImageButton
+                android:id="@+id/backBtn"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:adjustViewBounds="true"
@@ -27,6 +28,7 @@
             app:srcCompat="@drawable/btn_back" />
 
         <TextView
+                android:id="@+id/name"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_gravity="center_vertical"
@@ -38,6 +40,7 @@
             android:textSize="26sp" />
 
         <ImageButton
+                android:id="@+id/settingsBtn"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:adjustViewBounds="true"
