+++ app/src/main/res/layout/fragment_add_place.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    tools:context=".AddPlaceDialogFragment">
+    tools:context=".main.AddPlaceDialogFragment">
 
     <!-- 제목 -->
     <LinearLayout
@@ -111,7 +111,7 @@
     </LinearLayout>
 
     <androidx.fragment.app.FragmentContainerView
-            android:name="org.konkuk.placelist.MapFragment"
+            android:name="org.konkuk.placelist.main.MapFragment"
             android:id="@+id/map_fragment"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
